Refactor: move CSV import view to a separate module
