+++ repondeur/zam_repondeur/views/lectures.py
@@ -1,10 +1,8 @@
-import csv
-import io
 import logging
 import transaction
 from collections import Counter
 from datetime import datetime
-from typing import BinaryIO, Dict, Optional, TextIO, Tuple, Union
+from typing import BinaryIO, Dict, Optional, Union
 
 import ujson as json
 from pyramid.httpexceptions import HTTPBadRequest, HTTPFound, HTTPNotFound
@@ -13,7 +11,6 @@ from pyramid.response import Response
 from pyramid.view import view_config, view_defaults
 from sqlalchemy.orm import joinedload
 
-from zam_repondeur.clean import clean_html
 from zam_repondeur.data import get_data
 from zam_repondeur.fetch.an.dossiers.models import Dossier, Lecture
 from zam_repondeur.message import Message
@@ -25,10 +22,6 @@ from zam_repondeur.resources import (
 )
 from zam_repondeur.tasks.fetch import fetch_articles, fetch_amendements
 from zam_repondeur.utils import normalize_avis, normalize_num, normalize_reponse
-
-
-class CSVError(Exception):
-    pass
 
 
 @view_config(context=LectureCollection, renderer="lectures_list.html")
@@ -183,105 +176,6 @@ class ListAmendements:
         }
 
 
-@view_config(context=LectureResource, name="import_csv", request_method="POST")
-def import_csv(context: LectureResource, request: Request) -> Response:
-
-    lecture = context.model()
-
-    next_url = request.resource_url(context["amendements"])
-
-    # We cannot just do `if not POST["reponses"]`, as FieldStorage does not want
-    # to be cast to a boolean.
-    if request.POST["reponses"] == b"":
-        request.session.flash(
-            Message(cls="warning", text="Veuillez d’abord sélectionner un fichier")
-        )
-        return HTTPFound(location=next_url)
-
-    try:
-        reponses_count, errors_count = _import_reponses_from_csv_file(
-            reponses_file=request.POST["reponses"].file,
-            amendements={
-                amendement.num: amendement for amendement in lecture.amendements
-            },
-        )
-    except CSVError as exc:
-        request.session.flash(Message(cls="danger", text=str(exc)))
-        return HTTPFound(location=next_url)
-
-    if reponses_count:
-        request.session.flash(
-            Message(
-                cls="success",
-                text=f"{reponses_count} réponse(s) chargée(s) avec succès",
-            )
-        )
-        lecture.modified_at = datetime.utcnow()
-
-    if errors_count:
-        request.session.flash(
-            Message(
-                cls="warning",
-                text=(
-                    f"{errors_count} réponse(s) n’ont pas pu être chargée(s). "
-                    "Pour rappel, il faut que le fichier CSV contienne au moins "
-                    "les noms de colonnes suivants « Num amdt », "
-                    "« Avis du Gouvernement », « Objet amdt » et « Réponse »."
-                ),
-            )
-        )
-
-    return HTTPFound(location=next_url)
-
-
-def _import_reponses_from_csv_file(
-    reponses_file: BinaryIO, amendements: Dict[int, Amendement]
-) -> Tuple[int, int]:
-    previous_reponse = ""
-    reponses_count = 0
-    errors_count = 0
-
-    reponses_text_file = io.TextIOWrapper(reponses_file, encoding="utf-8-sig")
-
-    delimiter = _guess_csv_delimiter(reponses_text_file)
-
-    for line in csv.DictReader(reponses_text_file, delimiter=delimiter):
-        try:
-            numero = line["Num amdt"]
-            avis = line["Avis du Gouvernement"] or ""
-            objet = line["Objet amdt"] or ""
-            reponse = line["Réponse"] or ""
-        except KeyError:
-            errors_count += 1
-            continue
-
-        try:
-            num = normalize_num(numero)
-        except ValueError:
-            logging.warning("Invalid amendement number %r", numero)
-            errors_count += 1
-            continue
-
-        amendement = amendements.get(num)
-        if not amendement:
-            logging.warning("Could not find amendement number %r", num)
-            errors_count += 1
-            continue
-
-        amendement.avis = normalize_avis(avis)
-        amendement.observations = clean_html(objet)
-        reponse = normalize_reponse(reponse, previous_reponse)
-        amendement.reponse = clean_html(reponse)
-        if "Affectation" in line:
-            amendement.affectation = clean_html(line["Affectation"])
-        if "Commentaires" in line:
-            amendement.comments = clean_html(line["Commentaires"])
-        previous_reponse = reponse
-        reponses_count += 1
-
-    return reponses_count, errors_count
-
-
 @view_config(context=LectureResource, name="import_backup", request_method="POST")
 def import_backup(context: LectureResource, request: Request) -> Response:
 
@@ -397,27 +291,6 @@ def _import_backup_from_json_file(
         counter["articles"] += 1
 
     return counter
-
-
-def _guess_csv_delimiter(text_file: TextIO) -> str:
-    try:
-        sample = text_file.readline()
-    except UnicodeDecodeError:
-        raise CSVError("Le fichier n’est pas encodé en UTF-8")
-    except Exception:
-        raise CSVError("Le format du fichier n’est pas reconnu")
-
-    try:
-        dialect = csv.Sniffer().sniff(sample, delimiters=",;\t")
-    except csv.Error:
-        raise CSVError(
-            "Le fichier CSV n’utilise pas un délimiteur reconnu "
-            "(virgule, point-virgule ou tabulation)"
-        )
-
-    text_file.seek(0)
-
-    return dialect.delimiter
 
 
 @view_config(context=LectureResource, name="manual_refresh")
